+++ gazpacho/__init__.py
@@ -1,2 +1,2 @@
-from .soup import Soup
 from .get import get
+from .soup import Soup
